+++ src/main/java/core/Currency/RegalOrb.java
@@ -13,7 +13,7 @@ import core.Utils.AddRandomMod;
 
 public class RegalOrb implements Crafting_Action {
 	
-	private double cost = 1.0;
+	private int cost = 1;
 	
 	public enum CurrencyTier {
 		BASE, GREATER, PERFECT
@@ -101,7 +101,7 @@ public class RegalOrb implements Crafting_Action {
 	}
 
 	@Override
-	public double getCost() {
+	public int getCost() {
 		return cost;
 	}
